// src/modules/auth/types/userGBS.type.ts
/* ––
 * –––– Interface declaration
 * –––––––––––––––––––––––––––––––––– */
export default interface UserGBS {
  address: string;
  agentId: string;
  bookId: string;
  city: string;
  country: string;
  customerBirthDate: string;
  email: string;
  firstName: string;
  key: string;
  lastName: string;
  loginID: string;
  password: string;
  phone: string;
  promoCode: string;
  referredBy: string;
  state: string;
  zip: string;
}